refactor: split core logic
