+++ README.md
@@ -24,6 +24,7 @@
 * v0.3.0: 麻雀牌譜ログ構造とエクスポート機能追加
 * v0.4.0: シミュレーション結果をCSVダウンロード可能に
 * v0.5.0: ESLint導入とCIでのチェックを追加
+* v0.6.0: コア機能とUI部分を分離
 
 ## 主な機能
